==> feed/4.1_livestock_diet_total_consumption.Rmd <==
---
title: "Calculating regional totals for feedstuff"
author: "Juliette"
date: "10/21/2020"
output: html_document
editor_options: 
  chunk_output_type: console
--- 

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(here)
source(here("_workflow/common.R"))
library(stringi)


## files we'll need
## feed consumption rates
pigs_feedstuff_rate <- read_csv(here("feed/data/pigs_feed_consumption_rates.csv"))
chickens_feedstuff_rate <- read_csv(here("feed/data/chicken_feed_consumption_rates.csv"))
ruminants_feedstuff_rate <- read_csv(here("feed/data/ruminant_feed_consumption_rates.csv"))

```

Need a list of all the counts, but the mixed and feedlot counts need to come from the separate files
```{r}

rgn_counts_list  <- list.files(file.path(prep, "animal_farm/farm/rgn_df"), pattern = "region_counts", full = TRUE)

```


```{r}

combine_feedstuff <- data_frame(iso3c = NA, animal = NA, system = NA, product = NA, feedstuff = NA, kg_feedstuff = NA, tonnes_feedstuff = NA)


for(file in rgn_counts_list) {

    # file <- rgn_counts_list[6]
    animal_name <- str_split(str_split(file, pattern = "/")[[1]][12], pattern = "_")[[1]][3]
    system_name <- str_split(str_split(file, pattern = "/")[[1]][12], pattern = "_")[[1]][4]
    product_name <- str_split(str_split(str_split(file, pattern = "/")[[1]][12], pattern = "_")[[1]][5], pattern = ".csv")[[1]][1]
    
    rgn_counts <- read_csv(file)
    
    if(animal_name %in% c("cows", "sheep", "goats", "buffaloes")) {
  
      feedstuff_rate <- read_csv(here("feed/data/ruminant_feed_consumption_rates.csv"))
   
       }else{
         
         if(animal_name == "chickens"){
           feedstuff_rate <- read_csv(here("feed/data/chicken_feed_consumption_rates.csv"))
           
           }else{
             feedstuff_rate <- read_csv(here("feed/data/pigs_feed_consumption_rates.csv"))
           }}
      
      feedstuff_rate <- feedstuff_rate %>% 
        filter(animal == animal_name,
               product == product_name,
               system == system_name)
      
      total_feedstuff <- left_join(feedstuff_rate, rgn_counts, by = c("iso3c", "animal", "system", "product")) %>% 
        mutate(kg_feedstuff = rgn_count*feeditem_kg_per_head,
               tonnes_feedstuff = kg_feedstuff/1000) %>% 
        select(iso3c, animal, system, product, feedstuff, kg_feedstuff, tonnes_feedstuff)
        
    
  combine_feedstuff <- rbind(total_feedstuff, combine_feedstuff) %>% 
    na.omit()
      
  write_csv(combine_feedstuff, here("feed/data/total_livestock_feedstuff_consumption.csv"))    

}


```

FOFM are overestimated in chickens and pigs.  Making a correction so the final values are consistent with Froehlich et al.  2018.


Check that fishoil/fishmeal matches Halley data.
Specifically matching to Fig.3 in paper, data here: https://knb.ecoinformatics.org/view/doi:10.5063/F1T43R7N
(New_Current_Taxon_Fig.csv (for figures))
```{r}

combine_feedstuff <- read_csv(here("feed/data/total_livestock_feedstuff_consumption.csv"))    

halley_tonnes <- read.csv(here("feed/data/msleckman.45.1_fofm_consumption.csv")) %>%
  filter(!is.na(animal)) %>%
  group_by(animal) %>%
  summarize(true_tonnes_fish = sum(Current_Mean)) %>%
  mutate(true_tonnes_fofm = true_tonnes_fish*0.28)  # 0.28 converts tonnes fish to fofm

fofm <- combine_feedstuff %>%
    filter(feedstuff == "FISHMEAL") %>%
    group_by(animal, feedstuff) %>%
  summarize(tonnes_fofm=sum(tonnes_feedstuff, na.rm=TRUE)) %>% 
  left_join(halley_tonnes) %>%
  mutate(correction = true_tonnes_fofm/tonnes_fofm)

write_csv(fofm, here("feed/data/fofm_livestock_corrected_consumption.csv"))
```



```{r}

combine_feedstuff_corrected <- combine_feedstuff %>%
 left_join(fofm) %>%
  mutate(correction = ifelse(is.na(correction), 1, correction)) %>%
  mutate(kg_feedstuff_fofm_correct = kg_feedstuff*correction) %>%
  mutate(tonnes_feedstuff_fofm_correct = tonnes_feedstuff*correction) %>%
  select(iso3c, animal, system, product, feedstuff, kg_feedstuff=kg_feedstuff_fofm_correct, tonnes_feedstuff=tonnes_feedstuff_fofm_correct)

write_csv(combine_feedstuff_corrected, here("feed/data/total_livestock_feedstuff_consumption.csv"))    

# check; Halley values and ours should be similar (comparing "total" which is the fofm estimate from "true_tonnes_fish" with true_tonnes_fofm
combine_feedstuff_corrected %>%
    group_by(animal, feedstuff) %>%
  summarize(total=sum(tonnes_feedstuff, na.rm=TRUE)) %>%    #
  filter(feedstuff == "FISHMEAL") %>%
  left_join(halley_tonnes)

# check that these are still the same (make sure only the fish stuff was adjusted)
combine_feedstuff_corrected %>%
    group_by(animal, feedstuff) %>%
  summarize(total=sum(tonnes_feedstuff, na.rm=TRUE)) %>%    
  filter(feedstuff == "Grains from maize (Zea mays)") 

```



Look at global feed breakdown 

```{r}
library(hrbrthemes)

feedstuff_raw <- read_csv(here("feed/data/total_livestock_feedstuff_consumption.csv")) 

feedstuff <- feedstuff_raw %>% 
  group_by(feedstuff) %>% 
  dplyr::summarise(sum_tonnes = sum(tonnes_feedstuff)) %>% 
  ungroup() %>% 
  arrange((sum_tonnes)) %>% 
  mutate(feedstuff = as_factor(feedstuff))

plot <- ggplot(feedstuff) +
  geom_col(aes(y = feedstuff, x = sum_tonnes, fill = feedstuff))+
#   theme_ipsum_rc() +
  theme(legend.position = "none",
        axis.text.y = element_text(size = 8))+
  
  labs(y = "Feedstuff", x = "Tonnes Feedstuff",
       title = "Global Livestock Feedstuff") 

plot 
ggsave(here("feed/plots/total_feed_plot_11.9.png"), width = 15, height = 10, dpi=300) 

```
