==> feed/6.2_feedfish_catch_proportion.Rmd <==
---
title: "Forage fish for feed"
author: "Melanie Frazier (UCSB, NCEAS, OHI)"
date: "May 21, 2019"
output: html_document
editor_options: 
  chunk_output_type: console
---

GOAL: Create a raster describing the proportion of each country's catch going to animal feed in the location of capture.

```{r}
#load relevant packages, etc.

library(here)
library(raster)
library(tidyverse)
library(countrycode)

# function to project raster
## Use correct CRS


watson_raster_template <- raster::raster(ncol=720, nrow=360, vals=c(1:259200))
food_raster <- raster(nrows=2160, ncols=4320, xmn=-180, xmx=180, ymn=-90, ymx=90)

source(here("fisheries/marine/marine_fisheries_old/raster_crs.R"))

```


The likely country source of fish meal/oil for each animal system.  These data were generated by Jessica Gephart based on country level consumption and trade data. 
```{r}

feed_rgn <- read_csv(here("feed/data/FMFO_bySource_2021Apr.csv")) %>%
  filter(!is.na(tonnes))

```

ISSUE 1:
There are some regions in these data that are not in the Watson catch data which will result in the loss of this catch.  In some cases, loss is substantial so we will account for this later and distribute globally. 

```{r}
# exploring what is lost

watson_rgn <- read_csv(here("fisheries/marine/ghg/data/Watson_rgns_with_iso3c.csv"))

excluded <- setdiff(feed_rgn$iso3c, watson_rgn$iso3c)

# catch for regions being lost
filter(feed_rgn, iso3c %in% excluded) %>%
  group_by(system) %>%
  summarize(total = sum(tonnes))

# total catch
feed_rgn %>%
  group_by(system) %>%
  summarize(total = sum(tonnes, na.rm=TRUE))

```


Determine total forage fish captured by each country (Watson catch data):
```{r}

foragefish <- read_csv("/home/shares/food-systems/Food_footprint/all_food_systems/dataprep/fisheries/marine/ghg/watson_v5_emissions/full_watson_catch_data_2017.csv")

foragefish_rgn_totals <- foragefish %>%
  group_by(iso3c) %>%
  summarize(FOFM_catch = sum(FOFM_catch))

```

Divide each country's total pelagic feed fish catch by the total going to each animal feed system.
```{r}

FOFM_props <- left_join(foragefish_rgn_totals, feed_rgn, by="iso3c") %>%
  filter(!is.na(system)) %>% # these regions have catch, but none recorded going to a system
  mutate(prop = ifelse(FOFM_catch == 0, 0, tonnes/FOFM_catch))

summary(FOFM_props)  

```

ISSUE 2: in some cases, we overshoot and say there are more fish caught in the country for feed than what is reported in the fisheries catch data.  

In these cases: we constrain the consumed catch to not overshoot the reported catch for a country and distribute this extra globally later on.

```{r}

# list of overshooting countries
FOFM_props %>%
  group_by(iso3c) %>%
  summarize(total_prop = sum(prop)) %>% 
  filter(total_prop>1)

# code to limit feed to available catch:
FOFM_props_limit <- FOFM_props %>%
  group_by(iso3c) %>%
  mutate(total_prop = sum(prop, na.rm=TRUE),
            total_tonnes = sum(tonnes)) %>%
  rowwise() %>%
  mutate(prop_consumption = tonnes/total_tonnes) %>%
  mutate(prop_correct = ifelse(total_prop>1, prop_consumption, prop)) %>%
  select(iso3c, country, FOFM_catch, system, tonnes, prop=prop_correct)
  
## peak and see that all looks fine
filter(FOFM_props_limit, iso3c=="SYC") %>% data.frame()
filter(FOFM_props_limit, iso3c=="USA") %>% data.frame()
  
```

All sources of loss result in about 15% loss of fofm catch.  This will be assigned to global catch proportional to total forage fish catch. We will need to calculate the amount of fofm catch lost, and add that to our total. 

```{r}

# total fofm catch:
forage_tonnes <- sum(foragefish_rgn_totals$FOFM_catch) # 39054068

# total system consumption:
system_consumption <- read_csv(here("feed/data/FMFO_country_data.csv")) %>%
  group_by(system) %>%
  summarize(total_feed = sum(tonnes, na.rm=TRUE)) %>%
  filter(!is.na(system))

loss <- FOFM_props_limit %>%
  rowwise() %>%
  mutate(est_tonnes = FOFM_catch*prop) %>%
  group_by(system) %>%
  summarize(est_feed = sum(est_tonnes)) %>%
  left_join(system_consumption, by="system") %>%
  mutate(prop_accounted_for = est_feed/total_feed) %>%
  mutate(loss = total_feed - est_feed) %>%
  mutate(forage_tonnes = forage_tonnes) %>%
  mutate(prop_of_global_distribute = loss/forage_tonnes)

# about 15% is being lost
sum(loss$est_feed, na.rm=TRUE)/sum(loss$total_feed, na.rm=TRUE) #0.0.85

```

create rasters describing proportion of each countries fofm catch going to feed for each food system category.

```{r}
systems <- loss$system

for(system_item in systems){ # system_item = systems[1]

  watson_raster_template <- raster::raster(ncol=720, nrow=360, vals=c(1:259200))

  
  system_prop <- FOFM_props_limit %>%
  dplyr::filter(system == system_item) %>%
  dplyr::select(iso3c, prop)

system_extra_prop <- filter(loss, system == system_item) %>%
  pull(prop_of_global_distribute)

foragefish_system <- left_join(foragefish, system_prop, by="iso3c") %>%
  mutate(prop = ifelse(is.na(prop), 0, prop)) %>% ## fix NAs
  mutate(system_extra_prop = system_extra_prop) %>%
  mutate(FOFM_catch_foodsystem_ID = FOFM_catch * prop) %>% ## fofm catch to feed
  mutate(FOFM_catch_foodsystem_xtra = FOFM_catch * system_extra_prop) %>% ## fofm catch to feed that is missing
  mutate(FOFM_catch_foodsystem = FOFM_catch_foodsystem_ID + FOFM_catch_foodsystem_xtra) %>% ## add that together
  group_by(Cell) %>%
  summarize(tonnes = sum(FOFM_catch_foodsystem, na.rm=TRUE))
  
forage_fish_system <- raster::subs(watson_raster_template, foragefish_system, by = "Cell", which = "tonnes", subsWithNA=TRUE)

# plot(log(forage_fish_system + 1))

# the following should be about equal
print(sprintf("the following should be about equal for: %s", system_item))
print(sprintf("raster calc: %s", cellStats(forage_fish_system, "sum", na.rm=TRUE))) 
print(sprintf("observed: %s", filter(loss, system==system_item) %>% pull(total_feed)))

# divide by total feed fish catch to get proportion
total_tonnes <- raster("/home/shares/food-systems/Food_footprint/all_food_systems/dataprep/fisheries/marine/ghg/watson_v5_emissions/catch_rasters/forage_fish_tonnes_2017.tif")

#cellStats(total_tonnes, "sum", na.rm=TRUE)
system_prop <- forage_fish_system/total_tonnes
#plot(system_prop)

# check that cell stats match before and after
#cellStats(system_prop, "mean", na.rm=TRUE) 
system_prop_crs <- projectRaster(system_prop, food_raster, method="ngb", over=TRUE)
#cellStats(system_prop_crs, "mean", na.rm=TRUE)

category1 = ifelse(grepl("chicken|pig|goat|cow|buffalo", system_item), "land", "marine")

writeRaster(system_prop_crs, sprintf("/home/shares/food-systems/Food_footprint/all_food_systems/datalayers/%s_%s_x_marine_fofm_fisheries_meat.tif", category1, system_item), overwrite=TRUE)

}

```

Check
```{r}

prop <- raster("/home/shares/food-systems/Food_footprint/all_food_systems/datalayers/marine_salmon_aquaculture_meat_x_marine_fofm_fisheries_meat.tif")
total_tonnes_new_crs <- raster("/home/shares/food-systems/Food_footprint/all_food_systems/dataprep/fisheries/marine/ghg/watson_v5_emissions/catch_rasters/forage_fish_tonnes_new_crs_2017.tif")
cellStats(total_tonnes_new_crs, "sum", na.rm=TRUE)

check <- prop*total_tonnes_new_crs
cellStats(check, "sum", na.rm=TRUE)

feed_rgn <- read_csv(here("feed/data/FMFO_bySource.csv"))
feed_rgn %>%
  group_by(system) %>%
  summarize(total = sum(tonnes, na.rm=TRUE))


2813887/2813892 # .9999982

```

## once all is calculated, do the extra consumption, which is mostly human:
```{r}
prop_feed_files <- list.files("/home/shares/food-systems/Food_footprint/all_food_systems/datalayers", pattern="_x_marine_fofm_fisheries_meat.tif", full=TRUE)

prop_feed_files <- grep("land_human_human_human_x", prop_feed_files, value=TRUE, invert=TRUE)

total_tonnes_new_crs <- raster("/home/shares/food-systems/Food_footprint/all_food_systems/dataprep/fisheries/marine/ghg/watson_v5_emissions/catch_rasters/forage_fish_tonnes_new_crs_2017.tif")

cellStats(total_tonnes_new_crs, "sum", na.rm=TRUE)
plot(total_tonnes_new_crs)

prop_feed_stack <- stack(prop_feed_files)

total_feed_props <- calc(prop_feed_stack, sum, na.rm=TRUE, progress = "text")

total_feed_props
plot(total_feed_props)
leftover_prop <- 1 - total_feed_props
plot(leftover_prop)
leftover_prop[leftover_prop < 0] <- 0
plot(leftover_prop)

writeRaster(leftover_prop, "/home/shares/food-systems/Food_footprint/all_food_systems/datalayers/land_human_human_human_x_marine_fofm_fisheries_meat.tif", overwrite=TRUE)

## including all the files including the food ones should result in values of 1 for all for fofm catch
prop_feed_files <- list.files("/home/shares/food-systems/Food_footprint/all_food_systems/datalayers", pattern="_x_marine_fofm_fisheries_meat.tif", full=TRUE)
prop_feed_stack <- stack(prop_feed_files)

total_feed_props <- calc(prop_feed_stack, sum, na.rm=TRUE, progress = "text")
total_feed_props
plot(total_feed_props)

new <- total_feed_props * total_tonnes_new_crs
plot(new)
```
